requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0

# Web interface
flask>=2.3.0
Werkzeug>=2.3.0

# AI/ML providers
google-generativeai>=0.3.0
openai>=1.0.0
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
matplotlib>=3.7.0

# Configuration management
PyYAML>=6.0.0

# Video processing
moviepy>=1.0.3

# Testing
pytest>=7.0.0

# Environment management
python-dotenv>=1.0.0

# Optional: For better performance
# numba>=0.57.0